Cleanup code from patch-multiquery, move UTF8 Interop code to extensions
