+++ src/SQLite.Net.Platform.Shared.Interop/SQLite.Net.Platform.Shared.Interop.projitems
@@ -9,6 +9,7 @@
     <Import_RootNamespace>SQLite.Net.Platform.Shared.Interop</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)InteropExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)SQLite3.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)SQLite3Configuration.cs" />
   </ItemGroup>
